Move lsp implementation details to nls
