+++ lsp/nls/src/requests/mod.rs
@@ -2,4 +2,3 @@ pub mod completion;
 pub mod goto;
 pub mod hover;
 pub mod symbols;
-mod utils;
